Update SmartWord.java
I added a method to load the words into the Trie. I also added a method to print all the words in the trie to make sure it's getting everything. I also commented out some of the methods so that I could test this, so just uncomment those if you want to work on those methods.